AI-8410: fix v2 sidebar CSS and finish layout trim
Always inject default sidebar styles in v2 shell (match core PR).
styleTheme only adds WordPress CSS. Drop SDK toggle injection,
inline preset defaults, and narrow openEmbeddedIframe args.
