--- src/load-sidebar-v2/presets/sidebar.ts
@@ -1,7 +1,0 @@
-export const SIDEBAR_PRESET_DEFAULTS = {
-	layout: 'sidebar' as const,
-	styleTheme: '' as const,
-	persistOpenState: true,
-	resizable: true,
-	chatToggleButtonEnabled: false,
-} as const;
