+++ src/load-sidebar-v2/sidebar-toggle.ts
@@ -1,11 +1,9 @@
-import { findToggleButton, injectToggleButton } from './toggle-button';
-
 type WireSidebarToggleButtonArgs = {
 	toggleButtonSelector: string;
 };
 
 export const syncSidebarToggleButton = ( toggleButtonSelector: string, isOpen: boolean ): void => {
-	const toggleEl = findToggleButton( toggleButtonSelector );
+	const toggleEl = document.querySelector<HTMLElement>( toggleButtonSelector );
 
 	if ( ! toggleEl ) {
 		return;
@@ -30,31 +28,12 @@ const attachSidebarToggleClickHandler = ( toggleEl: HTMLElement ): void => {
 	} );
 };
 
-export const wireSidebarToggleButton = ( args: WireSidebarToggleButtonArgs ): boolean => {
-	const toggleEl = findToggleButton( args.toggleButtonSelector );
+export const wireSidebarToggleButton = ( args: WireSidebarToggleButtonArgs ): void => {
+	const toggleEl = document.querySelector<HTMLElement>( args.toggleButtonSelector );
 
 	if ( ! toggleEl ) {
-		return false;
+		return;
 	}
 
 	attachSidebarToggleClickHandler( toggleEl );
-	return true;
-};
-
-export const ensureSidebarToggleButton = ( args: WireSidebarToggleButtonArgs ): void => {
-	const attempt = (): void => {
-		if ( wireSidebarToggleButton( args ) ) {
-			return;
-		}
-
-		injectToggleButton( args.toggleButtonSelector );
-
-		if ( wireSidebarToggleButton( args ) ) {
-			return;
-		}
-
-		setTimeout( attempt, 500 );
-	};
-
-	setTimeout( attempt, 100 );
 };
